Improve Spectra selection and display Queue status
Improved the Spectra selection in model.py to avoid errors when the spectra dataframe is empty. Superfluous white spaces and line breaks were also removed for better code readability. The current queue status is now displayed on the home.py. Commented out the water_loss and ammonia_loss parameters in the utils.py as these are not available in current version of peptacular.